moodtune_lmd/requirements.txt
httpx
pyYAML
pytest
pytest-asyncio
fastapi
uvicorn
jinja2
pydantic
boto3
